[DS-707] Correctness fixes
git-svn-id: http://scm.dspace.org/svn/repo/dspace/trunk@5560 9c30dcfa-912a-0410-8fc2-9e0234be79fd

+++ dspace-api/src/main/java/org/dspace/app/itemexport/ItemExport.java
@@ -846,7 +846,7 @@ public class ItemExport
         // keep track of the commulative size of all bitstreams in each of the
         // items
         // it will be checked against the config file entry
-        float size = 0;
+        double size = 0;
         final ArrayList<Integer> items = new ArrayList<Integer>();
         for (DSpaceObject dso : dsObjects)
         {
